+++ R/loglinear_interpolation.R
@@ -5,7 +5,8 @@ NULL
 #'
 #' Log-linearly interpolate values between two thresholds.
 #'
-#' @param x \code{numeric} \emph{x} values for which interpolate emph{y} values.
+#' @param x \code{numeric} \emph{x} values for which interpolate \emph{y}
+#'   values.
 #'
 #' @param coordinate_one_x \code{numeric} value for lower \emph{x}-coordinate.
 #'
@@ -16,7 +17,7 @@ NULL
 #' @param coordinate_two_y \code{numeric} value for upper \emph{y}-coordinate.
 #'
 #' @details Values are log-linearly interpolated at the \emph{x}-coordinates
-#'   specified in \code{x} using the lower and upper coordnate arguments to
+#'   specified in \code{x} using the lower and upper coordinate arguments to
 #'   define the line. Values lesser or greater than these numbers are assigned
 #'   the minimum and maximum \emph{y} coordinates.
 #'