+++ R/marxan_boundary_data_to_matrix.R
@@ -14,7 +14,7 @@ NULL
 #' @param x \code{\link{ConservationProblem-class}} object that contains
 #'   planning unit and zone data to ensure that the argument to
 #'   \code{data} is converted correctly. This argument can be set to
-#'   \code{NULL} if checks are not required (not recomended).
+#'   \code{NULL} if checks are not required (not recommended).
 #'
 #' @param data \code{data.frame} object with the columns \code{"id1"},
 #'   \code{"id2"}, and \code{"boundary"}. The columns \code{"zone1"} and