+++ R/simulate.R
@@ -67,7 +67,7 @@ simulate_data <- function(x, n, model, transform = identity, ...) {
   return(stk)
 }
 
-#' Simulate species habitat suitabilities
+#' Simulate species habitat suitability data
 #'
 #' Generates a random set of species using random field models. By default,
 #' the output will contain values between zero and one.